restructured with Next.js
rebuild projcet structure with Next.js
