+++ .eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "extends": "next/core-web-vitals"
+  // "rules": {
+  //   "react/display-name": "off"
+  // }
+}
