--- .prettierrc
@@ -1,6 +1,0 @@
-{
-  "arrowParens": "avoid",
-  "trailingComma": "es5",
-  "singleQuote": true,
-  "semi": true
-}
