+++ components/contact.jsx
@@ -159,4 +159,6 @@ const Contact = forwardRef((props, ref) => {
     );
 });
 
+Contact.displayName = 'Contact';
+
 export default Contact;
