+++ next.config.js
@@ -1,0 +1,16 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+
+}
+
+module.exports = nextConfig
+
+// const withImages = require('next-images')
+// module.exports = withImages()
+
+// module.exports = {  
+//   images: {
+//     disableStaticImage: true,
+//   },
+// };
