+++ package.json
@@ -1,60 +1,39 @@
 {
   "name": "jk_portfolio_v2",
   "version": "0.1.0",
-  "homepage": "http://Mark-JunHyung-Kim.github.io/JK_Portfolio_v2",
   "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint",
+    "vercel-deploy": "next build && next export"
+  },
   "dependencies": {
     "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.5",
+    "@chakra-ui/react": "^1.8.6",
     "@emailjs/browser": "^3.4.0",
-    "@emotion/react": "^11.8.1",
+    "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
-    "@fontsource/open-sans": "^4.5.5",
+    "@fontsource/open-sans": "^4.5.6",
     "@fontsource/raleway": "^4.5.3",
     "@fortawesome/fontawesome-svg-core": "^1.3.0",
     "@fortawesome/free-brands-svg-icons": "^6.0.0",
     "@fortawesome/free-regular-svg-icons": "^6.0.0",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
     "@fortawesome/react-fontawesome": "^0.1.17",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^10.4.9",
-    "@testing-library/user-event": "^12.8.3",
     "formik": "^2.2.9",
     "formik-chakra-ui": "^1.6.0",
-    "framer-motion": "^4.1.17",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^3.11.0",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
+    "framer-motion": "^6.2.8",
+    "next": "12.1.0",
+    "next-images": "^1.8.4",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "swiper": "^8.0.7",
-    "web-vitals": "^0.2.4",
     "yup": "^0.32.11"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
-    "gh-pages": "^3.2.3"
+    "eslint": "8.11.0",
+    "eslint-config-next": "12.1.0"
   }
 }
