+++ styles/index.jsx
@@ -3,9 +3,8 @@ import { extendTheme } from '@chakra-ui/react'
 const theme = extendTheme({
   fonts: {
     heading: 'Open Sans, sans-serif',
-    body: 'Raleway, sans-serif'    
+    body: 'Raleway, sans-serif',
   },
-  
 })
 
 export default theme
